Unify exception handlers
